drm/exynos: rewrite userptr support
The current userptr support is broken. This rewrites the support
introducing a new ioctl to issue userptr operations.

userptr now have to be registered and unregistered explicitly.
Registered userptrs can then be used in the G2D command list
as read/write targets.

This breaks the existing API but since the support never
worked in the first place, this can be safely ignored.